Cleanup, add Zoomify viewer if not enough control points, updated README.

+++ public/javascripts/maphub/TileOverlay.js
@@ -14,16 +14,36 @@ maphub.TileOverlay = function(parameters) {
 	if (parameters) {
 		this.map = parameters.map;
 		this.tileSize = parameters.tileSize;
+		this.url = this.map.metadata.overlay_tileset_uri;
 		
+		/*
+		 * Get a reference to ourselves for use in callbacks.
+		 */
 		var self = this;
-		google.maps.event.addListener(this.map.googleMap, 'idle', function () {
-			self.deleteHiddenTiles();
-		});
+		
+		/*
+		 * If requested, enable the deletion of tiles that aren't in the viewer.
+		 */
+		if (this.map['deleteHiddenTiles']) {
+			google.maps.event.addListener(this.map.googleMap, 'idle', function () {
+				self.deleteHiddenTiles();
+			});
+		}
 	}
 }
 
+
+
+/*
+ * Delete tiles that are not visible in the viewport. This helps keep memory
+ * overhead low, which is noticable on wimpy systems like phones and implanted
+ * tracking microchi--er--just phones.
+ */
 maphub.TileOverlay.prototype.deleteHiddenTiles = function() {
-	console.log("Deleting hidden tiles.");
+//	console.log("Deleting hidden tiles.");
+	/*
+	 * First, set some variables to extract the boundaries of the current map.
+	 */
 	var bounds = this.map.googleMap.getBounds();
 	var zoom = this.map.googleMap.getZoom();
 	var ne = bounds.getNorthEast();
@@ -32,29 +52,23 @@ maphub.TileOverlay.prototype.deleteHiddenTiles = function() {
 	var swTile = this.latLngToTile(sw, zoom);
 	var nwTile = new google.maps.Point(swTile.x, neTile.y);
 	var seTile = new google.maps.Point(neTile.x, swTile.y);
-	console.log("New tiles: "+neTile+","+swTile+","+nwTile+","+seTile);
-	
+//	console.log("New tiles: "+neTile+","+swTile+","+nwTile+","+seTile);
+
+	/*
+	 * Determine the boundaries of the viewport in terms of tile coordinates.
+	 */
 	var tileNE = this.getTileUrlCoordFromLatLng(bounds.getNorthEast(), zoom);
 	var tileSW = this.getTileUrlCoordFromLatLng(bounds.getSouthWest(), zoom);
-//	var minX = tileSW.x - 1;
-//	var maxX = tileNE.x + 1;
-//	var minY = tileSW.y - 1;
-//	var maxY = tileNE.y + 1;
 	var minX = tileSW.x;
 	var maxX = tileNE.x;
 	var minY = tileSW.y;
 	var maxY = tileNE.y;
-	
-	
-	console.log('Should be (14,10) to (21,12)');
-	console.log('('+minX+','+minY+') to ('+maxX+','+maxY+')');
 
 	/*
-	 * Create an object in which we will store the tiles we want to keep. This will eventually
-	 * be assigned to this.tiles.
+	 * Create an object in which we will store the tiles we want to keep. This
+	 * will eventually be assigned to this.tiles.
 	 */
 	var tilesToKeep = {};
-	console.log("Before: "+Object.keys(this.tiles));
 	
 	/*
 	 * Loop through all current tiles.
@@ -67,7 +81,7 @@ maphub.TileOverlay.prototype.deleteHiddenTiles = function() {
 			for (var y = minY; y <= maxY; y++) {
 				var tmpTileID = zoom+'-'+x+'-'+y;
 				if (tmpTileID == tileID) {
-					console.log(tileID+" matches");
+//					console.log(tileID+" matches");
 					/*
 					 * This tile is visible, keep it.
 					 */
@@ -78,37 +92,23 @@ maphub.TileOverlay.prototype.deleteHiddenTiles = function() {
 	}
 	
 	this.tiles = tilesToKeep;
-	console.log("After: "+Object.keys(this.tiles));
 	return;
-	
-	for (var i = 0; i < tilesLength; i++) {
-		var idParts = this.tiles[i].id.split("_");
-		var tileX = Number(idParts[1]);
-		var tileY = Number(idParts[2]);
-		var tileZ = Number(idParts[3]);
-		if ((
-				(minX < maxX && (tileX >= minX && tileX <= maxX))
-				|| (minX > maxX && ((tileX >= minX && tileX <= (Math.pow(2, zoom) - 1)) || (tileX >= 0 && tileX <= maxX))) // Lapped the earth!
-			)
-			&& (tileY >= minY && tileY <= maxY)
-			&& tileZ == zoom) {
-			tilesToKeep.push(this.tiles[i]);
-		}
-		else {
-			delete this.tiles[i];
-		}
-	}
-	
-	this.tiles = tilesToKeep;
 }
 
+
+
+/*
+ * Retrieve the specified tile and creater a DIV element in the specified
+ * container.
+ *
+ * @param point The tile coordinate from Google, which doesn't match up with the
+ * 	overlay map's coordinates due to different projections.
+ * @param zoomLevel The current zoom level.
+ * @param container The container in which the tile should be created.
+ * @return The DIV element that was created.
+ */
 maphub.TileOverlay.prototype.getTile = function(point, zoomLevel, container) {
 //	console.log("maphub.TileOverlay.prototype.getTile");
-
-	var xynw = new google.maps.Point(point.x*this.tileSize.width, point.y*this.tileSize.height);
-	var xyse = new google.maps.Point((point.x+1)*this.tileSize.width, (point.y+1)*this.tileSize.height);
-	
-//	console.log("Coord for "+point+': '+xynw+', '+xyse);
 	
 	/*
 	 * If the tile already exists, use it.
@@ -116,41 +116,55 @@ maphub.TileOverlay.prototype.getTile = function(point, zoomLevel, container) {
 	var tileID = zoomLevel + '-' + point.x + '-' + point.y;
 	var div = null;
 	if (typeof this.tiles[tileID] == "undefined") {
-		/*
-		 * The tile doesn't exist, create it.
-		 */
 //		console.log("Creating new tile for " + tileID);
+		/*
+		 * The tile doesn't exist, create it. First, find out what the map's
+		 * tile coordinates (not lat/lng or x/y) are. Since the map uses a flat
+		 * x,y projection, and Google uses a spherical Mercator projection, we
+		 * need to use the map's metadata to find out which tile is the correct
+		 * tile to use for the specified Google map tile.
+		 */
+ 		var tileCoordinates = this.getTileUrlCoord(point, zoomLevel);
+
+		/*
+		 * Second, create the actual HTML element and set the background image to
+		 * the tile's image.
+		 */
 		div = container.createElement('div');
 		div.style.width = this.tileSize.width + 'px';
 		div.style.height = this.tileSize.height + 'px';
+		div.style.background = 'url('+this.url+'/'+zoomLevel+'/'+tileCoordinates.x+'/'+tileCoordinates.y+'.png) no-repeat';
+//		div.style.border = '1px solid #888';
+//		div.innerHTML = '('+point.x+','+point.y+')';
 
-//		var ymax = 1 << zoomLevel;
-//		var y = ymax - point.y - 1;
-//		div.style.background = 'url(/images/tiles/ct002033/google/'+zoomLevel+'/'+point.x+'/'+y+'.png) no-repeat';
-		var tileCoordinates = this.getTileUrlCoord(point, zoomLevel);
-		div.style.background = 'url(/images/tiles/'+this.map.id+'/google/'+zoomLevel+'/'+tileCoordinates.x+'/'+tileCoordinates.y+'.png) no-repeat';
-
-		div.style.border = '1px solid #888';
-		div.innerHTML = '('+point.x+','+point.y+')';
-		
+		/*
+		 * Cache the tile for quick retrieval later.
+		 */
 		this.tiles[tileID] = div;
 	} else {
 		/*
 		 * The tile already exists, use it.
 		 */
-		console.log("Found existing tile for " + tileID);
+//		console.log("Found existing tile for " + tileID);
 		div = this.tiles[tileID];
 	}
 
 	return div;
 }
 
+
+
+/*
+ * Convert the specified lat/lng pair and zoom level to a tile coordinate. This
+ * takes into account the fact that the MapHub map is using a different (flat)
+ * projection than the Google map (Mercator).
+ */
 maphub.TileOverlay.prototype.latLngToTile = function(latLng, z) {
 	var projection = this.map.googleMap.getProjection();
 	var worldCoordinate = projection.fromLatLngToPoint(latLng);
 	var pixelCoordinate = new google.maps.Point(worldCoordinate.x * Math.pow(2, z), worldCoordinate.y * Math.pow(2, z));
 	var tileCoordinate = new google.maps.Point(Math.floor(pixelCoordinate.x / this.tileSize.width), Math.floor(pixelCoordinate.y / this.tileSize.height));
-	console.log("lLTT: "+worldCoordinate+" "+pixelCoordinate+" "+tileCoordinate);
+//	console.log("lLTT: "+worldCoordinate+" "+pixelCoordinate+" "+tileCoordinate);
 	return tileCoordinate;
 }
 
@@ -160,7 +174,7 @@ maphub.TileOverlay.prototype.getTileUrlCoordFromLatLng = function (latlng, zoom)
 }
 
 maphub.TileOverlay.prototype.getTileUrlCoord = function(coord, zoom) {
-	console.log('maphub.TileOverlay.prototype.getTileUrlCoord: '+coord+', '+zoom);
+//	console.log('maphub.TileOverlay.prototype.getTileUrlCoord: '+coord+', '+zoom);
 	var tileRange = 1 << zoom;
 	var y = tileRange - coord.y - 1;
 	var x = coord.x;
